ai-api/requirements.txt
openai
python-dotenv
elevenlabs
langchain
llama-index
argparse
duckduckgo-search

huggingface_hub
tiktoken
wikipedia
transformers


redis
uvicorn
fastapi
pydantic